+++ src/Channel/Changelog.md
@@ -8,6 +8,7 @@
 - Dropped PHP 8.0 & PHP 8.1 Support
 - Dropped Laravel 9 Support
 - Added PHP 8.3 Support
+- Added Laravel 11 Support
 - Changed minimum Laravel version to v10.38.2
 - Added the following fields to the Channel model/table:
   - `currency`