+++ src/Channel/Changelog.md
@@ -10,7 +10,7 @@
 - Added PHP 8.3 Support
 - Added Laravel 11 Support
 - Added the vanilo/support dependency
-- Changed minimum Laravel version to v10.38.2
+- Changed minimum Laravel version to v10.43
 - Added the following fields to the Channel model/table:
   - `currency`
   - `language`